job-scout-graphic-design/requirements.txt
requests
beautifulsoup4
python-dotenv
pandas
html5lib
lxml
